Update to use a dynamic programming approach (#10)
If we use a message pass for every combinator, the parser is really
slow. If we use recursive descent, it uses too much memory, because
nothing is getting garbage collected. This PR makes the parser use a
dynamic programming approach that does a limited amount of processing in
one behaviour, but then calls the behaviour again so as to enable
garbage collection.